refactor(keepcommon): dedupe subscription package, point generator at pkg/subscription
pkg/keepcommon/subscription was byte-for-byte identical to the
existing pkg/subscription. Delete the duplicate, repoint the vendored
generator template and already-generated contract bindings at the
single remaining copy.
