+++ tools/generators/ethereum/contract.go.tmpl
@@ -19,7 +19,7 @@ import (
 
 	chainutil "{{.ChainUtilPackage}}"
 	"github.com/keep-network/keep-core/pkg/keepcommon/chain/ethereum"
-	"github.com/keep-network/keep-core/pkg/keepcommon/subscription"
+	"github.com/keep-network/keep-core/pkg/subscription"
 {{if .GenPackage}}
 	"{{.GenPackage}}/abi"
 {{end}}
